+++ packages/components/src/index.js
@@ -19,3 +19,4 @@ export { default as InstructionsList } from './instructionsList';
 export { default as ProcessIndicator } from './processIndicator';
 export { default as Upload } from './upload';
 export { default as DefinitionList } from './definitionList';
+export { default as Button } from './button';